+++ src/services/token/storage/index.ts
@@ -1,4 +1,4 @@
-import { Token } from '../authorizer';
+import { Token } from '../../authorizer';
 
 export interface ITokenStorage {
   save(token: Token): Promise<void>;
